// packages/shared/logic/core/Types.ts
export type BlueprintType =
    | 'METEOR_SHOWER'
    | 'NEURAL_OVERCLOCK'
    | 'STASIS_FIELD'
    | 'PERK_RESONANCE'
    | 'ARENA_SURGE'
    | 'QUANTUM_SCRAPPER'
    | 'MATRIX_OVERDRIVE'
    | 'TEMPORAL_GUARD'
    | 'DIMENSIONAL_GATE'
    | 'SECTOR_UPGRADE_ECO'
    | 'SECTOR_UPGRADE_COM'
    | 'SECTOR_UPGRADE_DEF';

export interface Blueprint {
    id: number;
    type: BlueprintType;
    name: string;
    serial: string;
    desc: string;
    cost: number;
    duration: number;
    isBlueprint: boolean;
    researched: boolean;
    status: 'locked' | 'ready' | 'active' | 'broken' | 'researching';
    discoveredIn?: string;
    researchRemainingTime?: number;
    researchFinishTime?: number;
    researchDuration?: number;
}